Update assembler.py
It now builds a list of legal opcodes to check against. (This list may not be 100% correct, not going to exhaustively test it right this second.

Also, the assembler now produces more useful errors when it fails.
The error messages include the line number and the line that caused the error.